pyproject.toml
[project]
name = "streaming-workshop"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "kafka-python>=2.3.0",
    "pandas>=3.0.1",
    "psycopg2-binary>=2.9.11",
    "pyarrow>=23.0.1",
    "requests>=2.32.5",
]

[dependency-groups]
dev = [
    "jupyter>=1.1.1",
]
